refactor: split profile schema modules
