--- extension/src/lib/schema.ts
@@ -1,300 +1,0 @@
-import { uuidv7 } from "uuidv7";
-import { z } from "zod";
-import { PROFILE_IMPORT_SCHEMA_VERSION } from "./const";
-
-const uuidSchemaWithDefault = z.uuid().default(() => uuidv7());
-const groupItemSchema = z.object({
-  enabled: z.boolean(),
-  comments: z.string().optional(),
-  id: uuidSchemaWithDefault,
-});
-export type GroupItem = z.infer<typeof groupItemSchema>;
-
-const groupItemWithValueSchema = groupItemSchema.extend({
-  value: z.string(),
-});
-
-const redirectUrlGroupItemSchema = groupItemWithValueSchema;
-export type RedirectUrlGroupItem = z.infer<typeof redirectUrlGroupItemSchema>;
-
-const groupTypeSchema = z.enum(["radio", "checkbox"]);
-export type GroupType = z.infer<typeof groupTypeSchema>;
-
-function createFilterGroupSchema<T extends z.ZodType>(itemSchema: T) {
-  return z.object({
-    type: groupTypeSchema,
-    items: z.array(itemSchema),
-  });
-}
-
-export const profileGroupSchema = z.object({
-  id: uuidSchemaWithDefault,
-  name: z.string(),
-  color: z.string(),
-  type: groupTypeSchema,
-  lastEnabledProfileIds: z.array(z.uuid()).optional(),
-});
-export type ProfileGroup = z.infer<typeof profileGroupSchema>;
-
-export const RESOURCE_TYPES = [
-  "main_frame",
-  "sub_frame",
-  "stylesheet",
-  "script",
-  "image",
-  "font",
-  "object",
-  "xmlhttprequest",
-  "ping",
-  "csp_report",
-  "media",
-  "websocket",
-  "webtransport",
-  "webbundle",
-  "other",
-] as const satisfies `${Browser.declarativeNetRequest.ResourceType}`[];
-const resourceTypeSchema = z.enum(RESOURCE_TYPES);
-export type ResourceType = z.infer<typeof resourceTypeSchema>;
-
-export const REQUEST_METHODS = [
-  "connect",
-  "delete",
-  "get",
-  "head",
-  "options",
-  "patch",
-  "post",
-  "put",
-  "other",
-] as const satisfies `${Browser.declarativeNetRequest.RequestMethod}`[];
-const requestMethodSchema = z.enum(REQUEST_METHODS);
-
-export const RULE_ACTION_TYPES = [
-  "block",
-  "redirect",
-  "allow",
-  "upgradeScheme",
-  "modifyHeaders",
-  "allowAllRequests",
-] as const satisfies `${Browser.declarativeNetRequest.RuleActionType}`[];
-
-const ruleActionTypeSchema = z.enum(RULE_ACTION_TYPES);
-export type RuleActionType = z.infer<typeof ruleActionTypeSchema>;
-
-export const DOMAIN_TYPES = [
-  "firstParty",
-  "thirdParty",
-] as const satisfies `${Browser.declarativeNetRequest.DomainType}`[];
-const domainTypeValueSchema = z.enum(DOMAIN_TYPES);
-
-const removeOperationSchema = groupItemSchema.extend({
-  name: z.string(),
-  operation: z.literal("remove"),
-});
-
-const appendOrSetOperationSchema = groupItemSchema.extend({
-  name: z.string(),
-  operation: z.enum(["append", "set"]),
-  value: z.string(),
-});
-
-const headerModSchema = z.union([
-  removeOperationSchema,
-  appendOrSetOperationSchema,
-]);
-export type HeaderMod = z.infer<typeof headerModSchema>;
-
-const headerModGroupSchema = z.object({
-  id: uuidSchemaWithDefault,
-  type: groupTypeSchema,
-  items: z.array(headerModSchema),
-});
-export type HeaderModGroup = z.infer<typeof headerModGroupSchema>;
-
-const urlOrRegexFilterSchema = groupItemWithValueSchema;
-export type UrlOrRegexFilterItem = z.infer<typeof urlOrRegexFilterSchema>;
-
-const domainsFilterSchema = createFilterGroupSchema(groupItemWithValueSchema);
-export type DomainsFilter = z.infer<typeof domainsFilterSchema>;
-
-const resourceTypesFilterSchema = groupItemSchema.extend({
-  value: z.array(resourceTypeSchema),
-});
-
-const resourceTypesFilterGroupSchema = createFilterGroupSchema(resourceTypesFilterSchema);
-export type ResourceTypesFilterGroup = z.infer<typeof resourceTypesFilterGroupSchema>;
-
-const requestMethodsFilterSchema = groupItemSchema.extend({
-  value: z.array(requestMethodSchema),
-});
-
-const requestMethodsFilterGroupSchema = createFilterGroupSchema(requestMethodsFilterSchema);
-export type RequestMethodsFilterGroup = z.infer<typeof requestMethodsFilterGroupSchema>;
-
-const tabIdsFilterSchema = groupItemSchema.extend({
-  value: z.array(z.int().nonnegative()),
-});
-export type TabIdsFilterItem = z.infer<typeof tabIdsFilterSchema>;
-
-const tabIdsFilterGroupSchema = createFilterGroupSchema(tabIdsFilterSchema);
-export type TabIdsFilterGroup = z.infer<typeof tabIdsFilterGroupSchema>;
-
-const tabGroupBindingSchema = z.object({ groupId: z.int().nonnegative(), tabIds: z.array(z.int().nonnegative()) });
-export type TabGroupBinding = z.infer<typeof tabGroupBindingSchema>;
-const tabGroupsFilterSchema = groupItemSchema.extend({ value: z.array(tabGroupBindingSchema) });
-export type TabGroupsFilterItem = z.infer<typeof tabGroupsFilterSchema>;
-const tabGroupsFilterGroupSchema = createFilterGroupSchema(tabGroupsFilterSchema);
-export type TabGroupsFilterGroup = z.infer<typeof tabGroupsFilterGroupSchema>;
-const domainTypeFilterSchema = z.object({
-  enabled: z.boolean(),
-  value: domainTypeValueSchema,
-});
-
-const urlFilterCaseSensitiveSchema = z.object({
-  enabled: z.boolean(),
-  value: z.boolean(),
-});
-
-const syncCookieSchema = groupItemSchema.extend({
-  domain: z.string(),
-  name: z.string(),
-  value: z.string(),
-  path: z.string(),
-});
-export type SyncCookie = z.infer<typeof syncCookieSchema>;
-
-const syncCookieGroupSchema = z.object({
-  id: uuidSchemaWithDefault,
-  type: groupTypeSchema,
-  items: z.array(syncCookieSchema),
-});
-export type SyncCookieGroup = z.infer<typeof syncCookieGroupSchema>;
-
-const filterSchema = z.object({
-  urlFilter: z.array(urlOrRegexFilterSchema).optional(),
-  regexFilter: z.array(urlOrRegexFilterSchema).optional(),
-  initiatorDomains: domainsFilterSchema.optional(),
-  excludedInitiatorDomains: domainsFilterSchema.optional(),
-  requestDomains: domainsFilterSchema.optional(),
-  excludedRequestDomains: domainsFilterSchema.optional(),
-  topDomains: domainsFilterSchema.optional(),
-  excludedTopDomains: domainsFilterSchema.optional(),
-  resourceTypes: resourceTypesFilterGroupSchema.optional(),
-  excludedResourceTypes: resourceTypesFilterGroupSchema.optional(),
-  requestMethods: requestMethodsFilterGroupSchema.optional(),
-  excludedRequestMethods: requestMethodsFilterGroupSchema.optional(),
-  tabIds: tabIdsFilterGroupSchema.optional(),
-  excludedTabIds: tabIdsFilterGroupSchema.optional(),
-  tabGroups: tabGroupsFilterGroupSchema.optional(),
-  excludedTabGroups: tabGroupsFilterGroupSchema.optional(),
-  domainType: domainTypeFilterSchema.optional(),
-  isUrlFilterCaseSensitive: urlFilterCaseSensitiveSchema.optional(),
-});
-export type Filter = z.infer<typeof filterSchema>;
-
-const profileSchema = z.object({
-  id: uuidSchemaWithDefault,
-  name: z.string(),
-  enabled: z.boolean(),
-  emoji: z.string(),
-  groupId: z.uuid().optional(),
-  comments: z.string().optional(),
-  ruleActionType: ruleActionTypeSchema,
-  priority: z.int().min(1).optional().meta({ description: "Range: 1 to 2^31 - 1, default: 1" }),
-  requestHeaderModGroups: z.array(headerModGroupSchema).optional(),
-  responseHeaderModGroups: z.array(headerModGroupSchema).optional(),
-  syncCookieGroups: z.array(syncCookieGroupSchema).optional(),
-  redirectUrlGroup: z.array(redirectUrlGroupItemSchema).optional(),
-  filters: filterSchema,
-});
-export type Profile = z.infer<typeof profileSchema>;
-
-const removeOperationWithoutIdSchema = removeOperationSchema.omit({ id: true });
-const appendOrSetOperationWithoutIdSchema = appendOrSetOperationSchema.omit({ id: true });
-
-const headerModWithoutIdSchema = z.union([
-  removeOperationWithoutIdSchema,
-  appendOrSetOperationWithoutIdSchema,
-]);
-
-const headerModGroupWithoutIdSchema = headerModGroupSchema.omit({ id: true }).extend({
-  items: z.array(headerModWithoutIdSchema),
-});
-
-const syncCookieWithoutIdSchema = syncCookieSchema.omit({ id: true });
-
-const syncCookieGroupWithoutIdSchema = syncCookieGroupSchema.omit({ id: true }).extend({
-  items: z.array(syncCookieWithoutIdSchema),
-});
-
-const redirectUrlGroupItemWithoutIdSchema = redirectUrlGroupItemSchema.omit({ id: true });
-
-const urlOrRegexFilterWithoutIdSchema = urlOrRegexFilterSchema.omit({ id: true });
-
-const domainsFilterWithoutIdSchema = domainsFilterSchema.extend({
-  items: z.array(groupItemWithValueSchema.omit({ id: true })),
-});
-
-const resourceTypesFilterWithoutIdSchema = resourceTypesFilterSchema.omit({ id: true });
-const requestMethodsFilterWithoutIdSchema = requestMethodsFilterSchema.omit({ id: true });
-
-const resourceTypesFilterGroupWithoutIdSchema = createFilterGroupSchema(resourceTypesFilterWithoutIdSchema);
-const requestMethodsFilterGroupWithoutIdSchema = createFilterGroupSchema(requestMethodsFilterWithoutIdSchema);
-
-const filterWithoutIdSchema = filterSchema.extend({
-  urlFilter: z.array(urlOrRegexFilterWithoutIdSchema).optional(),
-  regexFilter: z.array(urlOrRegexFilterWithoutIdSchema).optional(),
-  initiatorDomains: domainsFilterWithoutIdSchema.optional(),
-  excludedInitiatorDomains: domainsFilterWithoutIdSchema.optional(),
-  requestDomains: domainsFilterWithoutIdSchema.optional(),
-  excludedRequestDomains: domainsFilterWithoutIdSchema.optional(),
-  topDomains: domainsFilterWithoutIdSchema.optional(),
-  excludedTopDomains: domainsFilterWithoutIdSchema.optional(),
-  resourceTypes: resourceTypesFilterGroupWithoutIdSchema.optional(),
-  excludedResourceTypes: resourceTypesFilterGroupWithoutIdSchema.optional(),
-  requestMethods: requestMethodsFilterGroupWithoutIdSchema.optional(),
-  excludedRequestMethods: requestMethodsFilterGroupWithoutIdSchema.optional(),
-}).omit({ tabIds: true, excludedTabIds: true, tabGroups: true, excludedTabGroups: true });
-
-export const profileWithoutIdsZodSchema = profileSchema.omit({ groupId: true, id: true }).extend({
-  requestHeaderModGroups: z.array(headerModGroupWithoutIdSchema).optional(),
-  responseHeaderModGroups: z.array(headerModGroupWithoutIdSchema).optional(),
-  syncCookieGroups: z.array(syncCookieGroupWithoutIdSchema).optional(),
-  redirectUrlGroup: z.array(redirectUrlGroupItemWithoutIdSchema).optional(),
-  filters: filterWithoutIdSchema,
-});
-
-export type ProfileWithoutIds = z.infer<typeof profileWithoutIdsZodSchema>;
-
-const profilesWithoutIdsArrayZodSchema = z.array(profileWithoutIdsZodSchema).min(1);
-
-export const profileExchangeZodSchema = z.object({
-  version: z.literal(PROFILE_IMPORT_SCHEMA_VERSION),
-  profiles: profilesWithoutIdsArrayZodSchema,
-});
-export type ProfileExchange = z.infer<typeof profileExchangeZodSchema>;
-
-export const profileExchangeJsonSchema = z.toJSONSchema(profileExchangeZodSchema);
-
-/**
- * Strip all id fields from profile
- * Uses zod parse feature to automatically strip fields not defined in schema (including all ids)
- */
-export function stripProfileIds(profile: Profile) {
-  return profileWithoutIdsZodSchema.parse(profile);
-}
-
-export function createProfileExchange(profiles: Profile[]): ProfileExchange {
-  return profileExchangeZodSchema.parse({
-    version: PROFILE_IMPORT_SCHEMA_VERSION,
-    profiles: profiles.map(stripProfileIds),
-  });
-}
-
-/**
- * Add id fields to profile
- * Uses z.default() in schema to automatically generate new UUIDs
- */
-export function addProfileIds(profileWithoutIds: ProfileWithoutIds) {
-  return profileSchema.parse(profileWithoutIds);
-}
