+++ extension/src/lib/schema/index.ts
@@ -1,0 +1,16 @@
+export {
+  DOMAIN_TYPES,
+  profileGroupSchema,
+  REQUEST_METHODS,
+  RESOURCE_TYPES,
+  RULE_ACTION_TYPES,
+} from "./base";
+export {
+  addProfileIds,
+  createProfileExchange,
+  profileExchangeJsonSchema,
+  profileExchangeZodSchema,
+  profileWithoutIdsZodSchema,
+  stripProfileIds,
+} from "./profile";
+export type * from "./type";
